kernel-fitimage: Add signature timestamp property
In fact mkimage is supposed to set the timestamp by itself when we
sign the image up, and one succeeds in doing this, but only if just
one image of each type defined. When at least two for instance flat_dt's
nodes are defined, then it fails with the next errors:
>> Can't set 'timestamp' property for 'signature@1' node (FDT_ERR_NOSPACE)
>> Can't write signature for 'signature@1' signature node in
>>      'conf@baikal_tplatforms_msbt2.dtb' conf node: <unknown error>
>> mkimage Can't add hashes to FIT blob

It seems as if mkimage can't allocate enough space for the signature node
while setting the timestamp for it. The reason might be in a next image
of the same type, which might block it.

The best way would be to fix the bug in the mkimage itself. But in any
case this workaround shall work for both fixed and unfixed versions of
mkimage.

Signed-off-by: Serge Semin <fancer.lancer@gmail.com>